+++ TablePro/Views/Sidebar/DatabaseTreeOutlineCoordinator+Commands.swift
@@ -149,14 +149,8 @@ extension DatabaseTreeOutlineCoordinator {
             ClipboardService.shared.writeText(key)
         case .openRedisKey(let key, let keyType):
             mainCoordinator?.openRedisKey(key, keyType: keyType)
-        case .toggleObjectIcons:
-            AppSettingsManager.shared.general.showObjectIcons.toggle()
-            refreshVisibleRows()
-        case .toggleObjectComments:
-            AppSettingsManager.shared.general.showObjectComments.toggle()
-            refreshVisibleRows()
-        case .setRowSize(let size):
-            AppSettingsManager.shared.general.sidebarRowSize = size
+        case .toggleObjectIcons, .toggleObjectComments, .setRowSize:
+            _ = SidebarViewOptionsMenu.apply(command)
         }
     }